Start to splitting up Vulkan monolith
